+++ Source/ACE/Entity/Creature.cs
@@ -89,11 +89,6 @@ namespace ACE.Entity
 
         public double RespawnTime { get; set; }
 
-        public Creature(AceObject aceObject, ObjectGuid guid)
-            : base(aceObject, guid)
-        {
-        }
-
         protected void SetupVitals()
         {
             if (Health.Current != Health.MaxValue)
@@ -113,6 +108,11 @@ namespace ACE.Entity
         public Creature(AceObject baseObject)
             : base(baseObject)
         {
+            if (Attackable == false)
+            {
+                if (RadarColor != Enum.RadarColor.Yellow || (RadarColor == Enum.RadarColor.Yellow && CreatureType == null))
+                        NpcLooksLikeObject = true;
+            }
         }
 
         public virtual ActionChain GetOnKillChain(Session killerSession)
@@ -501,12 +501,16 @@ namespace ACE.Entity
 
         public override void SerializeIdentifyObjectResponse(BinaryWriter writer, bool success, IdentifyResponseFlags flags = IdentifyResponseFlags.None)
         {
-            flags |= IdentifyResponseFlags.CreatureProfile;
+            bool hideCreatureProfile = NpcLooksLikeObject ?? false;
+
+            if (!hideCreatureProfile)
+            {
+                flags |= IdentifyResponseFlags.CreatureProfile;
+            }
 
             base.SerializeIdentifyObjectResponse(writer, success, flags);
 
-            // TODO: There are probably other checks that need to be made here
-            if (GetType().Name != "DebugObject")
+            if (!hideCreatureProfile)
             {
                 WriteIdentifyObjectCreatureProfile(writer, this, success);
             }