+++ Source/ACE/Factories/WorldObjectFactory.cs
@@ -27,9 +27,9 @@ namespace ACE.Factories
 
                 if (aceO.Location != null)
                 {
-                    ////WorldObject wo = new GenericObject(aceO);
-
-                    results.Add(CreateWorldObject(aceO));
+                    WorldObject wo = CreateWorldObject(aceO);
+                    if (wo != null)
+                        results.Add(wo);
                 }
             }
             return results;
@@ -38,17 +38,6 @@ namespace ACE.Factories
         public static WorldObject CreateWorldObject(AceObject aceO)
         {
             WeenieType objWeenieType = (WeenieType?)aceO.WeenieType ?? WeenieType.Generic;
-
-            ////if (aceO.GeneratorStatus ?? false)  // Generator
-            ////{
-            ////    ////aceO.Location = aceO.Location.InFrontOf(-2.0);
-            ////    ////aceO.Location.PositionZ = aceO.Location.PositionZ - 0.5f;
-            ////    ////results.Add(new Generator(new ObjectGuid(aceO.AceObjectId), aceO));
-            ////    ////aceO.GeneratorEnteredWorld = true;
-            ////    ////var objectList = GeneratorFactory.CreateWorldObjectsFromGenerator(aceO) ?? new List<WorldObject>();
-            ////    ////objectList.ForEach(o => results.Add(o));
-            ////    continue;
-            ////}
 
             switch (objWeenieType)
             {
@@ -61,7 +50,13 @@ namespace ACE.Factories
                 case WeenieType.Book:
                     return new Book(aceO);
                 // case WeenieType.PKModifier:
-                //    return new PKModifier(AceObject);
+                //    return new PKModifier(aceO);
+                case WeenieType.Cow:
+                    return new Cow(aceO);
+                case WeenieType.Creature:
+                    return new Creature(aceO);
+                case WeenieType.Container:
+                    return new Container(aceO);
                 default:
                     return new GenericObject(aceO);
             }
@@ -74,13 +69,23 @@ namespace ACE.Factories
             return CreateWorldObject(aceObject);
         }
 
+        public static WorldObject CreateWorldObject(uint weenieId, ObjectGuid guid)
+        {
+            AceObject aceObject = (AceObject)DatabaseManager.World.GetAceObjectByWeenie(weenieId).Clone(guid.Full);
+
+            return CreateWorldObject(aceObject);
+        }
+
         public static WorldObject CreateNewWorldObject(uint weenieId)
         {
-            // AceObject aceObject = DatabaseManager.World.GetAceObjectByWeenie(weenieId);
+            WorldObject wo = CreateWorldObject(weenieId, GuidManager.NewItemGuid());
 
-            WorldObject wo = CreateWorldObject(weenieId);
+            return wo;
+        }
 
-            wo.Guid = GuidManager.NewItemGuid(); // Assign new Guid
+        public static WorldObject CreateNewWorldObject(uint weenieId, ObjectGuid guid)
+        {
+            WorldObject wo = CreateWorldObject(weenieId, guid);
 
             return wo;
         }